+++ app/services/networkApi.js
@@ -1,6 +1,10 @@
 angular.module('uiForDocker')
   .service('networkApi', function($http) {
 		var networks  = {};
+    
+    networks.create = function(network) {
+      return $http.post('/api/networks', network);
+    };
 
     networks.getOne = function(id) {
       return $http.get('/api/networks/' + id);
@@ -13,6 +17,6 @@ angular.module('uiForDocker')
     networks.remove = function(id) {
       return $http.delete('/api/networks/' + id);
     };
-    
+
     return networks;
   });