+++ app/services/volumeApi.js
@@ -1,6 +1,10 @@
 angular.module('uiForDocker')
   .service('volumeApi', function($http) {
 		var volumes  = {};
+    
+    volumes.create = function(volume) {
+      return $http.post('/api/volumes', volume);
+    };
 
     volumes.getOne = function(id) {
       return $http.get('/api/volumes/' + id);